refactor: 프로필페이지 폴더이름 변경
나의 프로필만 조회하는 로직에서 모든 사용자의 프로필을 조회하는 로직으로 변경
