+++ src/pages/Profile/components/SignalTemperature/index.tsx
