+++ src/stories/pages/Profile/Review.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import MannerReview from '@/pages/MyProfile/components/MannerReview';
+import MannerReview from '@/pages/Profile/components/MannerReview';
 
 const meta: Meta<typeof MannerReview> = {
   title: 'Pages/MyProfile/MannerReview',
